refactor: Exception 패키지 경로 이동 & CustomException 을 상속 받도록 변경#5
