+++ src/main/java/com/cokkiri/secondhand/global/auth/service/GeneralAuthService.java
@@ -8,10 +8,10 @@ import com.cokkiri.secondhand.global.auth.dto.request.GeneralLogInRequest;
 import com.cokkiri.secondhand.global.auth.dto.request.GeneralSignUpRequest;
 import com.cokkiri.secondhand.global.auth.dto.response.JwtTokenResponse;
 import com.cokkiri.secondhand.global.auth.entity.UserInfoForJwt;
-import com.cokkiri.secondhand.global.exception.LoginFailureException;
-import com.cokkiri.secondhand.global.exception.NicknameDuplicationException;
-import com.cokkiri.secondhand.global.exception.NotExistLocationException;
-import com.cokkiri.secondhand.global.exception.UsernameDuplicationException;
+import com.cokkiri.secondhand.global.exception.list.LoginFailureException;
+import com.cokkiri.secondhand.global.exception.list.NicknameDuplicationException;
+import com.cokkiri.secondhand.global.exception.list.NotExistLocationException;
+import com.cokkiri.secondhand.global.exception.list.UsernameDuplicationException;
 import com.cokkiri.secondhand.item.entity.Location;
 import com.cokkiri.secondhand.item.repository.LocationJpaRepository;
 import com.cokkiri.secondhand.user.entity.GeneralUser;
