+++ src/main/java/com/cokkiri/secondhand/global/exception/CustomException.java
@@ -1,0 +1,8 @@
+package com.cokkiri.secondhand.global.exception;
+
+public class CustomException extends RuntimeException {
+
+	public CustomException(String message) {
+		super(message);
+	}
+}
