--- src/main/java/com/cokkiri/secondhand/global/exception/IllegalJwtTokenException.java
@@ -1,7 +1,0 @@
-package com.cokkiri.secondhand.global.exception;
-
-public class IllegalJwtTokenException extends RuntimeException {
-	public IllegalJwtTokenException(String message) {
-		super(message);
-	}
-}
