--- src/main/java/com/cokkiri/secondhand/global/exception/NicknameDuplicationException.java
@@ -1,8 +1,0 @@
-package com.cokkiri.secondhand.global.exception;
-
-public class NicknameDuplicationException extends RuntimeException {
-
-	public NicknameDuplicationException(String nickname) {
-		super(nickname + "은 이미 사용 중인 아이디 입니다.");
-	}
-}
