--- src/main/java/com/cokkiri/secondhand/global/exception/NotExistAccessTokenException.java
@@ -1,7 +1,0 @@
-package com.cokkiri.secondhand.global.exception;
-
-public class NotExistAccessTokenException extends RuntimeException {
-	public NotExistAccessTokenException() {
-		super("Access 토큰이 존재하지 않습니다.");
-	}
-}
