--- src/main/java/com/cokkiri/secondhand/global/exception/NotExistLocationException.java
@@ -1,7 +1,0 @@
-package com.cokkiri.secondhand.global.exception;
-
-public class NotExistLocationException  extends RuntimeException {
-	public NotExistLocationException(String message) {
-		super(String.format("[%s]는 존재하지 않는 위치 정보입니다.", message));
-	}
-}
