+++ src/main/java/com/cokkiri/secondhand/global/exception/list/IllegalJwtTokenException.java
@@ -1,0 +1,9 @@
+package com.cokkiri.secondhand.global.exception.list;
+
+import com.cokkiri.secondhand.global.exception.CustomException;
+
+public class IllegalJwtTokenException extends CustomException {
+	public IllegalJwtTokenException(String message) {
+		super(message);
+	}
+}
