+++ src/main/java/com/cokkiri/secondhand/global/exception/list/NotFoundUserByUsernameException.java
@@ -1,0 +1,10 @@
+package com.cokkiri.secondhand.global.exception.list;
+
+import com.cokkiri.secondhand.global.exception.CustomException;
+
+public class NotFoundUserByUsernameException extends CustomException {
+
+	public NotFoundUserByUsernameException(String nickname) {
+		super(nickname + "은 존재하지 않는 회원입니다.");
+	}
+}
